setup nwc to use new deeplinkhandler
